requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==24.1.1
certifi==2025.4.26
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
isort==5.13.2
jiter==0.9.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==1.26.4
openai==1.76.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
pyarrow==15.0.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
virtualenv==20.30.0
